move test out of sourc file
